# app/static/swagger.json
    {
    "openapi": "3.0.0",
    "info": {
        "title": "Job Portal API",
        "version": "1.0.0",
        "description": "구인구직 백엔드 서버 API 문서"
    },
    "servers": [
            {
                "url": "https://113.198.66.75:13062",
                "description": "Production server"
            },
            {
                "url": "http://localhost:3000",
                "description": "Development server"
            }
    ],
    "paths": {
...
        "/jobs": {
            "get": {
                "tags": ["Jobs"],
                "summary": "채용공고 목록 조회",
                "parameters": [
                    {
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "in": "query",
                        "name": "location",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "experience_level",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "min_salary",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "in": "query",
                        "name": "skills",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Comma separated skills"
                    },
                    {
                        "in": "query",
                        "name": "sort_by",
                        "schema": {
                            "type": "string",
                            "enum": ["salary", "deadline"]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "채용공고 목록",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "_id": {
                                                        "type": "string"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "company_name": {
                                                        "type": "string"
                                                    },
                                                    "location": {
                                                        "type": "string"
                                                    },
                                                    "salary_text": {
                                                        "type": "string"
                                                    },
                                                    "experience_level": {
                                                        "type": "string"
                                                    },
                                                    "education": {
                                                        "type": "string"
                                                    },
                                                    "job_type": {
                                                        "type": "string"
                                                    },
                                                    "skills": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string"
                                                        }
                                                    }
                                                }
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/Pagination"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/jobs/search": {
            "get": {
                "tags": ["Jobs"],
                "summary": "채용공고 검색",
                "parameters": [
                    {
                        "in": "query",
                        "name": "keyword",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "검색 결과",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/JobPosting"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/Pagination"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/jobs/{id}": {
            "get": {
                "tags": ["Jobs"],
                "summary": "채용공고 상세 조회",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "채용공고 상세 정보",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobPosting"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Pagination": {
                "type": "object",
                "properties": {
                    "currentPage": {
                        "type": "integer"
                    },
                    "totalPages": {
                        "type": "integer"
                    },
                    "totalItems": {
                        "type": "integer"
                    },
                    "perPage": {
                        "type": "integer",
                        "example": 20
                    }
                }
            },
            "JobPosting": {
                "type": "object",
                "properties": {
                        "_id": {
                            "type": "string"
                        },
                        "title": {
                            "type": "string"
                        },
                        "company_name": {
                            "type": "string"
                        },
                        "description": {
                            "type": "string"
                        },
                        "tasks": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        },
                        "requirements": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        },
                        "preferred": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        },
                        "benefits": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        },
                        "location": {
                            "type": "string"
                        },
                        "job_type": {
                            "type": "string"
                        },
                        "experience_level": {
                            "type": "string"
                        },
                        "education": {
                            "type": "string"
                        },
                        "salary_text": {
                            "type": "string"
                        },
                        "skills": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        },
                        "deadline": {
                            "type": "string"
                        },
                        "status": {
                            "type": "string",
                            "enum": ["active", "closed"]
                        },
                        "created_at": {
                            "type": "string",
                            "format": "date-time"
                        },
                        "updated_at": {
                            "type": "string",
                            "format": "date-time"
                        }
                    }
                }
            },
            "securitySchemes": {
                "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT Authorization header using the Bearer scheme. Example: 'Bearer {token}'"
                }
            },
...
        }
    }